Exit with a non-zero status when a task is not found
This process went through a few iterations. First, I tried adding a test
to the runner spec, but it seems that if you call exit in the src, it
will stop the test immediately so you can't actually check anything.

I ended up with this integration spec that creates a tasks.cr with a
single task inside of it. The test can now run tasks through that
instead of having to create a whole app skeleton.

I chose 127 as the exit code because it commonly means "command not
found". Technically, it is returning that code, but the number appears
to be 32512. This StackOverflow answer explains why:

https://stackoverflow.com/a/5638651/638966

The single, placeholder task is required because otherwise the task
runner doesn't have any tasks registered and therefore doesn't know the
type that `tasks_list` should return.