Remove Browsersync UI and make comments concise
The Browsersync UI is nice, but it also takes a lot of space to output
when first starting Lucky. There is already a lot there (too much), so
by default the UI is removed. If people want it they can add it later.

This also makes it easier/faster to jump to docs and removes unnecessary
language from the comments