Raise helpful permit columns error in Shared::Field
This error always happened when calling the input/label methods, but
if you used the Shared::Field component you'd get a generic "no overload
matches".

This module added in https://github.com/luckyframework/lucky/pull/1052
makes it so you get the same nice error in field components.

This greatly improves the ease of use and DX when using Lucky.