Disable 'Compiling' spinner on CI to avoid clogging logs
Resolves #560.

The Spinner class is complex, so I didn't really want to add to the complexity - especially with no tests. I considered implementing a `NullSpinner`, but the spinner interface is not super simple; it felt simpler to use the existing spinner class conditionally from its one usage in `BuildAndRunTask`.

It is curious that the spinner supports several unused features.